Move routing constraints to BopsCore engine
This is so they can be reused in other engines.
